fix: move calculations, update hook names [RWDQA-60]
