+++ src/components/annual-report/section4/SectionFour.js
@@ -1,8 +1,7 @@
 import PropTypes from 'prop-types'
 import React, { useEffect } from 'react'
 import { Chart } from '../Chart.js'
-import { calculateSection4 } from './section4Calculations.js'
-import { useFetchSectionFourData } from './useFetchSectionFourData.js'
+import { useSectionFourData } from './useSectionFourData.js'
 
 const sectionInformation = {
     section4a: {
@@ -135,7 +134,7 @@ Section4B.propTypes = {
 }
 
 export const SectionFour = ({ reportParameters }) => {
-    const { loading, data, error, refetch } = useFetchSectionFourData()
+    const { loading, data: section4Data, error, refetch } = useSectionFourData()
 
     useEffect(() => {
         const variables = {
@@ -153,14 +152,7 @@ export const SectionFour = ({ reportParameters }) => {
         return <span>error</span>
     }
 
-    if (data) {
-        const section4Data = calculateSection4({
-            section4Response: data.response,
-            mappedConfiguration: data.parameters?.mappedConfiguration,
-            currentPeriod: data.parameters?.periods?.[0],
-            overallOrgUnit: data.parameters?.orgUnits?.[0],
-        })
-
+    if (section4Data) {
         return (
             <>
                 <Section4A
